+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Removed `reconnect_delay_max` and `reconnect_delay`, you can pass precise sleep durations to `reconnect_attempts` instead.
 - Require Ruby 2.5+.
 - Removed the deprecated `queue` and `commit` methods. Use `pipelined` instead.
+- Removed the deprecated `Redis::Future#==`.
 - Removed the deprecated `pipelined` and `multi` signature. Commands now MUST be called on the block argument, not the original redis instance.
 - Removed `Redis.current`. You shouldn't assume there is a single global Redis connection, use a connection pool instead,
   and libaries using Redis should accept a Redis instance (or connection pool) as a config. E.g. `MyLibrary.redis = Redis.new(...)`.